requirements.txt
tk
dlib
opencv-python
Pillow
scipy
azure-cognitiveservices-vision-face
